Moved rap reader from utilities to io
